// BakeMyCake/src/app/Models/cakeOrderDetails.ts
export class CustomerDetails {
    id?: string;
    cakeName?: string;
    dateOfOrder?: string;
    customerName?: string;
    customerEmail?: string;
    customerPhone?: string;
    quantity?:number;
    street?:string;
    city?:string;
    state?:string;
    pinCode?:string;
    totalAmount?:number;
}